pyproject.toml
[tool.poetry]
name= "Alexandrya"
version= "0.1.0"
description= "App to manage library"
authors= ["Aranxa CODINA <aranxa.codina@aquarius.irish>"]

[tool.poetry.dependencies]
python= "3.8.*"
Django= "3.*"
django-cors-headers= "3.2.*"
django-environ = "^0.4.5"
django-extensions= "2.2.*"
django-ipware= "2.1.*"
django-oauth-toolkit= "1.3.*"
django-storages = {extras = ["azure"], version = "^1.9.1"}
djangorestframework= "3.11.*"
drf-nested-routers = "0.91.*"
gunicorn = "20.*"
mysqlclient = "1.4.*"
epub_meta = "^0.0.7"
dateutils = "^0.6.12"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
